pyproject.toml
[project]
name = "misbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Bohdan Stratila", email = "googlobody@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[webhooks]>=22.5",
    "uvicorn>=0.40.0",
]

[project.scripts]
server = "misbot:server"


[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
